requirements.txt
opencv-python
pillow~=9.0.1
tqdm
lxml
jupyter
matplotlib
pandas
scikit-learn
h5py
Cython
beautifulsoup4
requests~=2.27.1
selenium~=4.1.0